Separate esp-radio support code
